Moving new test into a separate file
