+++ packages/agreement-process/vitest.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   test: {
-    setupFiles: ["dotenv/config"],
+    setupFiles: ["dotenv/config", "test/agreementService.test.setup.ts"],
     testTimeout: 60000,
     hookTimeout: 60000,
   },
